完善比赛任务设置

+++ app/src/main/java/com/zhaoyan/ladderball/http/request/MatchDetailRequest.java
@@ -6,7 +6,7 @@ import android.content.Context;
  * Created by Yuri on 2015/12/7.
  */
 public class MatchDetailRequest extends BaseRequest{
-    public int matchId;
+    public long matchId;
 
     public MatchDetailRequest(Context context) {
         super(context);
